Move ml_model_base to unity/lib/extensions. Remove sdk_model_base.
Since the two model base classes are now identical, everything can just inherit from ml_model_base
