+++ src/unity/lib/extensions/ml_model.cpp
@@ -3,9 +3,10 @@
  * Use of this source code is governed by a BSD-3-clause license that can
  * be found in the LICENSE.txt file or at https://opensource.org/licenses/BSD-3-Clause
  */
+#include <unity/lib/extensions/ml_model.hpp>
+
 #include <fileio/general_fstream.hpp>
 #include <unity/lib/variant_deep_serialize.hpp>
-#include <unity/toolkits/ml_model/ml_model.hpp>
 
 // Unity global
 #include <unity/lib/unity_global.hpp>
