+++ src/unity/toolkits/ml_model/CMakeLists.txt
@@ -2,8 +2,6 @@ project(unity_toolkits)
 
 make_library(unity_ml_model
   SOURCES
-  ml_model.cpp
-  sdk_model.cpp
   python_model.cpp
   REQUIRES
     unity_core
