+++ src/unity/toolkits/nearest_neighbors/nearest_neighbors.hpp
@@ -21,7 +21,7 @@
 #include <unity/lib/toolkit_function_specification.hpp>
 #include <unity/lib/variant.hpp>
 #include <unity/lib/unity_base_types.hpp>
-#include <unity/toolkits/ml_model/ml_model.hpp>
+#include <unity/lib/extensions/ml_model.hpp>
 #include <unity/toolkits/util/algorithmic_utils.hpp>
 #include <toolkits/supervised_learning/supervised_learning_utils-inl.hpp>
